+++ CHANGES.rst
@@ -25,6 +25,9 @@ Changelog
   advisers when using inherited advices. This also fixes problem with optional
   key that was wrongly initialized for inherited advices.
   [gbastien]
+- Removed unused vocabularies `PMEveryCategoryVocabulary` and
+  `PMEveryCategoryTitleVocabulary`.
+  [gbastien]
 
 4.2.9rc2 (2024-02-26)
 ---------------------